jellyfin and jellyfin-ffmpeg:

+++ scripts/remove/jellyfin.sh
@@ -1,29 +1,38 @@
-#!/usr/bin/env bash
+#!/bin/bash
+. /etc/swizzin/sources/functions/utils
 #
-# Set the required variables
-username="$(cat /root/.master.info | cut -d: -f1)"
+username="$(_get_master_username)"
 #
-# Define the removal function for jellyfin.
-function remove_jellyfin() {
-    systemctl stop "jellyfin.service"
+systemctl -q stop "jellyfin.service"
+#
+if [[ -f /etc/systemd/system/jellyfin.service ]]; then
     #
-    systemctl disable "jellyfin.service"
-    #
+    ## legacy start
+    systemctl -q disable --now "jellyfin.service"
     rm -f "/etc/systemd/system/jellyfin.service"
-    #
     kill -9 $(ps xU ${username} | grep "/opt/jellyfin/jellyfin -d /home/${username}/.config/Jellyfin$" | awk '{print $1}') >/dev/null 2>&1
+    [[ -d "/opt/jellyfin" ]] && rm -rf "/opt/jellyfin" | :
+    [[ -d "/opt/ffmpeg" ]] && rm -rf "/opt/ffmpeg" | :
+    [[ -d "/home/${username}/.config/Jellyfin" ]] && rm -rf "/home/${username}/.config/Jellyfin" | :
+    [[ -d "/home/${username}/.cache/jellyfin" ]] && rm -rf "/home/${username}/.cache/jellyfin" | :
+    [[ -d "/home/${username}/.aspnet" ]] && rm -rf "/home/${username}/.aspnet" | :
+    ## legacy end
     #
-    rm -rf "/opt/jellyfin"
-    rm -rf "/opt/ffmpeg"
-    rm -rf "/home/${username}/.config/Jellyfin"
+else
+    apt_remove --purge jellyfin jellyfin-ffmpeg
     #
-    if [[ -f /install/.nginx.lock ]]; then
-        rm -f "/etc/nginx/apps/jellyfin.conf"
-        systemctl reload nginx
-    fi
-    #
-    rm -f "/install/.jellyfin.lock"
-}
+    [[ -d "/var/lib/jellyfin" ]] && rm -rf /var/lib/jellyfin | :
+    [[ -d "/var/log/jellyfin" ]] && rm -rf /var/log/jellyfin | :
+    [[ -d "/var/cache/jellyfin" ]] && rm -rf /var/cache/jellyfin | :
+    [[ -d "/usr/share/jellyfin/web" ]] && rm -rf /usr/share/jellyfin/web | :
+fi
 #
-# run the removal function
-remove_jellyfin
+# Remove the nginx conf and reload nginx.
+if [[ -f /install/.nginx.lock ]]; then 
+    rm -f /etc/nginx/apps/jellyfin.conf
+    systemctl -q reload nginx
+fi
+#
+rm -f /install/.jellyfin.lock
+#
+exit
